chore(ui): Move columns array inside the component
Move the columns array for the rule violations table, and for the
runs table inside the component in preparation for moving the
filtering to the table headers. This is because a `setFiltering`
function will be provided through the column meta field to the
header filter component, and it will need access to the useNavigate
hook.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>