feat: Return overall vulnerability rating for vulnerabilities of a run
Calculate the overall rating for the vulnerability from
the vulnerability references in the database query, and
return this rating in the response of the vulnerabilities
endpoint. This will allow for removing the calculation
from the frontend side, and any future improvements in
the calculation of the rating can be done in one place
on the backend side.

Resolves #1839.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>