fix(ui): Add handling for errors in the silent token renewal
Add a handler for errors during the silent renewal process. If the session
has expired, redirect to sign in straight away. Otherwise, in case of
network errors and such, retry the silent renewal after a timeout. Reuse
the existing `refreshUser` function for this, which will also handle any
further errors.

Resolves https://github.com/eclipse-apoapsis/ort-server/issues/2458.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>